ocr_logo_title/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pytesseract>=0.3.10
Pillow>=10.1.0
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.1.0
transformers>=4.35.0
python-multipart>=0.0.6
pydantic>=2.4.2 